Organize client modules under js directory
