+++ index.html
@@ -1487,6 +1487,6 @@
     </div>
 </section>
 
-<script src="app.js" type="module"></script>
+<script src="js/core/app.js" type="module"></script>
 </body>
 </html>
