+++ js/utils/audio.js
@@ -1,4 +1,4 @@
-import { AudioAssetPath } from "./constants.js";
+import { AudioAssetPath } from "../constants.js";
 
 // audio.js — web audio effects and buffered playback resources
